Jenkins-ci/jenkins-net-versionpackage-python/requirements.txt
python-jenkins==1.4.0
configparser==3.7.4
DingtalkChatbot==1.3.0
pytest-shutil==1.6.0
python-gitlab==1.8.0
zipfile36==0.1.3
